separated db functions, added delete CRUD API endpoint, updated read me to add CRUD table
