+++ README.md
@@ -1,6 +1,6 @@
 # Project Name
 
-> Project description
+> Optimizing database management system to support scaling the existing service
 
 ## Related Projects
 
@@ -11,9 +11,25 @@
 
 ## Table of Contents
 
+1. [CRUD](#CRUD)
 1. [Usage](#Usage)
 1. [Requirements](#requirements)
 1. [Development](#development)
+
+## CRUD
+
+> The below URL's should be prefixed with /api/products
+Action | Method | URL
+-------|--------|------
+Create a new question | POST | /:id/questions
+Create a new answer | POST | /:id/answers
+Get details for product | GET | /:id/details
+Get questions for product | GET | /:id/questions
+Get answers for product | GET | /:id/answers
+Update answer to Helpful | PUT | /:id/helpful
+Update answer to Not Helpful | PUT | /:id/not-helpful
+Delete a product | DELETE | /:id
+
 
 ## Usage
